01-Langchain_basico/pyproject.toml
[tool.poetry]
name = "b401"
version = "0.1.0"
description = ""
authors = ["julio colomer <info@aiaccelera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
jupyterlab = "^4.2.4"
langchain = "^0.2.10"
langchain-openai = "^0.1.17"
python-dotenv = "^1.0.1"
pypdf = "^6.8.0"
chromadb = "^1.5.5"
sentence-transformers = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
